### 2021-03-14 #3101 @whedon
[PRE REVIEW]: RustBCA: A High-Performance, Binary-Collision-Approximation Code for Ion-Material Interactions
**Submitting author:** @drobnyjt (<a href="http://orcid.org/0000-0002-9733-6058">Jon Drobny</a>)
**Repository:** <a href="https://github.com/lcpp-org/RustBCA" target ="_blank">https://github.com/lcpp-org/RustBCA</a>
**Version:** v1.0.0
