### 2021-03-15 7ddd574
Working parry3d::Ball example - runs with same input file as RustBCA's Sphere.
