+++ src/parry.rs
@@ -1,0 +1,135 @@
+use super::*;
+use parry3d::shape::{Ball};
+use parry3d::query::PointQuery;
+use parry3d::math::{Isometry, Point};
+
+#[derive(Deserialize, Clone)]
+pub struct InputParryBall {
+    pub options: Options,
+    pub material_parameters: material::MaterialParameters,
+    pub particle_parameters: particle::ParticleParameters,
+    pub geometry_input: ParryBallInput,
+}
+
+impl GeometryInput for InputParryBall {
+    type GeometryInput = ParryBallInput;
+}
+
+impl InputFile for InputParryBall {
+
+    fn new(string: &str) -> InputParryBall {
+        toml::from_str(string).expect("Could not parse TOML file.")
+    }
+
+    fn get_options(&self) -> &Options{
+        &self.options
+    }
+    fn get_material_parameters(&self) -> &material::MaterialParameters{
+        &self.material_parameters
+    }
+    fn get_particle_parameters(&self) ->& particle::ParticleParameters{
+        &self.particle_parameters
+    }
+    fn get_geometry_input(&self) -> &Self::GeometryInput{
+        &self.geometry_input
+    }
+}
+
+#[derive(Deserialize, Clone)]
+pub struct ParryBallInput {
+    pub length_unit: String,
+    pub radius: f64,
+    pub densities: Vec<f64>,
+    pub electronic_stopping_correction_factor: f64,
+}
+
+#[derive(Clone)]
+pub struct ParryBall {
+    pub densities: Vec<f64>,
+    pub concentrations: Vec<f64>,
+    pub radius: f64,
+    pub electronic_stopping_correction_factor: f64,
+    pub energy_barrier_thickness: f64,
+    pub ball: Ball,
+}
+
+impl GeometryInput for ParryBall {
+    type GeometryInput = ParryBallInput;
+}
+
+impl Geometry for ParryBall {
+
+    type InputFileFormat = InputParryBall;
+
+    fn new(input: &<Self as GeometryInput>::GeometryInput) -> ParryBall {
+
+        let length_unit: f64 = match input.length_unit.as_str() {
+            "MICRON" => MICRON,
+            "CM" => CM,
+            "ANGSTROM" => ANGSTROM,
+            "NM" => NM,
+            "M" => 1.,
+            _ => input.length_unit.parse()
+                .expect(format!(
+                        "Input errror: could nor parse length unit {}. Use a valid float or one of ANGSTROM, NM, MICRON, CM, MM, M",
+                        &input.length_unit.as_str()
+                    ).as_str()),
+        };
+
+        let electronic_stopping_correction_factor = input.electronic_stopping_correction_factor;
+        let densities: Vec<f64> = input.densities.iter().map(|element| element/(length_unit).powi(3)).collect();
+        let total_density: f64 = densities.iter().sum();
+        let energy_barrier_thickness = total_density.powf(-1./3.)/SQRTPI*2.;
+        let concentrations: Vec<f64> = densities.iter().map(|&density| density/total_density).collect::<Vec<f64>>();
+        let radius = input.radius*length_unit;
+
+        ParryBall {
+            densities,
+            concentrations,
+            radius,
+            electronic_stopping_correction_factor,
+            energy_barrier_thickness,
+            ball: Ball::new(radius as f32)
+        }
+    }
+
+    fn get_densities(&self,  x: f64, y: f64, z: f64) -> &Vec<f64> {
+        &self.densities
+    }
+    fn get_ck(&self,  x: f64, y: f64, z: f64) -> f64 {
[…]
