+++ src/enums.rs
@@ -5,6 +5,8 @@ pub enum MaterialType {
     MESH1D(material::Material<geometry::Mesh1D>),
     MESH2D(material::Material<geometry::Mesh2D>),
     SPHERE(material::Material<sphere::Sphere>),
+    #[cfg(feature = "parry")]
+    BALL(material::Material<parry::ParryBall>),
 }
 
 #[derive(Deserialize)]
@@ -13,6 +15,8 @@ pub enum GeometryType {
     MESH1D,
     MESH2D,
     SPHERE,
+    #[cfg(feature = "parry")]
+    BALL,
 }
 
 /// Mode of electronic stopping to use.
