+++ Cargo.toml
@@ -22,6 +22,7 @@ netlib-src = {version  = "0.8", optional = true}
 intel-mkl-src = {version = "0.6.0", optional = true}
 rcpr = { git = "https://github.com/drobnyjt/rcpr", optional = true}
 ndarray = {version = "0.14.0", features = ["serde"], optional = true}
+parry3d = {version = "0.2.0", optional = true}
 
 [dev-dependencies]
 float-cmp = "0.8.0"
@@ -39,3 +40,4 @@ cpr_rootfinder_netlib = ["rcpr", "netlib-src"]
 cpr_rootfinder_intel_mkl = ["rcpr", "intel-mkl-src"]
 distributions = ["ndarray"]
 no_list_output = []
+parry = ["parry3d"]
