### 2021-03-15 d64e19f
Working TriMesh input.
