+++ src/parry.rs
@@ -1,7 +1,7 @@
 use super::*;
-use parry3d::shape::{Ball};
-use parry3d::query::PointQuery;
-use parry3d::math::{Isometry, Point};
+use parry3d_f64::shape::{Ball, TriMesh};
+use parry3d_f64::query::{PointQuery, Ray, RayCast};
+use parry3d_f64::math::{Isometry, Point, Vector};
 
 #[derive(Deserialize, Clone)]
 pub struct InputParryBall {
@@ -89,7 +89,7 @@ impl Geometry for ParryBall {
             radius,
             electronic_stopping_correction_factor,
             energy_barrier_thickness,
-            ball: Ball::new(radius as f32)
+            ball: Ball::new(radius)
         }
     }
 
@@ -112,24 +112,197 @@ impl Geometry for ParryBall {
         self.electronic_stopping_correction_factor
     }
     fn inside(&self, x: f64, y: f64, z: f64) -> bool {
-        let p = Point::new(x as f32, y as f32, z as f32);
+        let p = Point::new(x , y , z );
         self.ball.contains_local_point(&p)
     }
 
     fn inside_simulation_boundary(&self, x: f64, y: f64, z: f64) -> bool {
-        let p = Point::new(x as f32, y as f32, z as f32);
+        let p = Point::new(x , y , z );
         (self.ball.distance_to_local_point(&p, true) as f64) < 10.*self.energy_barrier_thickness
     }
 
     fn inside_energy_barrier(&self, x: f64, y: f64, z: f64) -> bool {
-        let p = Point::new(x as f32, y as f32, z as f32);
+        let p = Point::new(x , y , z );
         (self.ball.distance_to_local_point(&p, true) as f64) < self.energy_barrier_thickness
     }
 
     fn closest_point(&self, x: f64, y: f64, z: f64) -> (f64, f64, f64) {
-        let p = Point::new(x as f32, y as f32, z as f32);
+        let p = Point::new(x , y , z );
         let point_projection = self.ball.project_point(&Isometry::identity(), &p, false);
         let (x_, y_, z_) = (point_projection.point.x, point_projection.point.y, point_projection.point.z);
         (x_ as f64, y_ as f64, z_ as f64)
+    }
+}
+
+
+#[derive(Deserialize, Clone)]
+pub struct InputParryTriMesh {
+    pub options: Options,
+    pub material_parameters: material::MaterialParameters,
+    pub particle_parameters: particle::ParticleParameters,
+    pub geometry_input: ParryTriMeshInput,
+}
+
+impl GeometryInput for InputParryTriMesh {
+    type GeometryInput = ParryTriMeshInput;
+}
+
+impl InputFile for InputParryTriMesh {
+
+    fn new(string: &str) -> InputParryTriMesh {
+        toml::from_str(string).expect("Could not parse TOML file.")
+    }
+
+    fn get_options(&self) -> &Options{
+        &self.options
+    }
+    fn get_material_parameters(&self) -> &material::MaterialParameters{
+        &self.material_parameters
+    }
+    fn get_particle_parameters(&self) ->& particle::ParticleParameters{
+        &self.particle_parameters
+    }
+    fn get_geometry_input(&self) -> &Self::GeometryInput{
+        &self.geometry_input
+    }
+}
+
+#[derive(Deserialize, Clone)]
+pub struct ParryTriMeshInput {
+    pub length_unit: String,
+    pub densities: Vec<f64>,
+    pub electronic_stopping_correction_factor: f64,
+    pub vertices: Vec<[f64; 3]>,
+    pub indices: Vec<[u32; 3]>,
+}
+
+#[derive(Clone)]
+pub struct ParryTriMesh {
+    pub densities: Vec<f64>,
+    pub concentrations: Vec<f64>,
[…]
