+++ examples/boron_nitride_trimesh.toml
@@ -58,4 +58,17 @@ vertices = [
   [-0.5, -0.5, -0.5],
   [0.5, -0.5, -0.5],
 ]
-indices =[[0, 1, 3], [0, 3, 2], [2, 3, 5], [2, 5, 4], [4, 5, 7], [4, 7, 6], [6, 7, 1], [6, 1, 0], [1, 7, 5], [1, 5, 3], [6, 0, 2], [6, 2, 4]]
+indices = [
+    [0, 1, 3],
+    [0, 3, 2],
+    [2, 3, 5],
+    [2, 5, 4],
+    [4, 5, 7],
+    [4, 7, 6],
+    [6, 7, 1],
+    [6, 1, 0],
+    [1, 7, 5],
+    [1, 5, 3],
+    [6, 0, 2],
+    [6, 2, 4]
+  ]
