### 2021-03-16 8492970
Updates to boundary calculation - now uses bounding box.
