### 2021-03-16 cf70397
Final changes to core geometry with parry3d.
