+++ examples/boron_nitride_trimesh.toml
@@ -1,74 +1,0 @@
-[options]
-name = "boron_nitride_"
-track_trajectories = true
-track_recoils = true
-track_recoil_trajectories = true
-write_buffer_size = 8000
-weak_collision_order = 0
-suppress_deep_recoils = false
-high_energy_free_flight_paths = false
-num_threads = 1
-num_chunks = 1
-use_hdf5 = false
-electronic_stopping_mode = "LOW_ENERGY_NONLOCAL"
-mean_free_path_model = "LIQUID"
-interaction_potential = [["KR_C"]]
-scattering_integral = [["MENDENHALL_WELLER"]]
-root_finder = [[{"NEWTON"={max_iterations=100, tolerance=1E-3}}]]
-track_displacements = false
-track_energy_losses = false
-
-[material_parameters]
-energy_unit = "EV"
-mass_unit = "AMU"
-Eb = [ 0.0, 0.0,]
-Es = [ 5.76, 0.0,]
-Ec = [ 1.0, 1.0,]
-Z = [ 5, 7,]
-m = [ 10.811, 14,]
-interaction_index = [0, 0]
-surface_binding_model = {"PLANAR"={calculation="TARGET"}}
-bulk_binding_model = "AVERAGE"
-
-[particle_parameters]
-length_unit = "1E-8"
-energy_unit = "EV"
-mass_unit = "AMU"
-N = [ 5 ]
-m = [ 1.008 ]
-Z = [ 1 ]
-E = [ 500.0 ]
-Ec = [ 1.0 ]
-Es = [ 10.0 ]
-pos = [ [ -0.5, 0.0, 0.0,] ]
-dir = [ [ 0.9999999999984769, 1.7453292519934434e-6, 0.0,] ]
-interaction_index = [ 0 ]
-
-[geometry_input]
-length_unit = "1E-8"
-electronic_stopping_correction_factor = 1.0
-densities = [65000, 65000]
-vertices = [
-  [-0.5, -0.5, 0.5],
-  [0.5, -0.5, 0.5],
-  [-0.5, 0.5, 0.5],
-  [0.5, 0.5, 0.5],
-  [-0.5, 0.5, -0.5],
-  [0.5, 0.5, -0.5],
-  [-0.5, -0.5, -0.5],
-  [0.5, -0.5, -0.5],
-]
-indices = [
-    [0, 1, 3],
-    [0, 3, 2],
-    [2, 3, 5],
-    [2, 5, 4],
-    [4, 5, 7],
-    [4, 7, 6],
-    [6, 7, 1],
-    [6, 1, 0],
-    [1, 7, 5],
-    [1, 5, 3],
-    [6, 0, 2],
-    [6, 2, 4]
-  ]
