+++ src/bca.rs
@@ -84,6 +84,13 @@ pub fn single_ion_bca<T: Geometry>(particle: particle::Particle, material: &mate
             //Choose impact parameters and azimuthal angles for all collisions, and determine mean free path
             let binary_collision_geometries = bca::determine_mfp_phi_impact_parameter(&mut particle_1, &material, &options);
 
+            #[cfg(feature = "accelerated_ions")]
+            if !material.inside(particle_1.pos.x, particle_1.pos.y, particle_1.pos.z) {
+                let (x, y, z) = material.geometry.closest_point(particle_1.pos.x, particle_1.pos.y, particle_1.pos.z);
+                let distance_to = ((x - particle_1.pos.x).powi(2) + (y - particle_1.pos.y).powi(2) + (z - particle_1.pos.z).powi(2)).sqrt();
+                mfp += distance_to;
+            }
+
             let mut total_energy_loss = 0.;
             let mut total_asymptotic_deflection = 0.;
             let mut normalized_distance_of_closest_approach = 0.;
@@ -330,11 +337,6 @@ pub fn determine_mfp_phi_impact_parameter<T: Geometry>(particle_1: &mut particle
             mfp *= -rand::random::<f64>().ln();
         }
 
-        if !material.inside(particle_1.pos.x, particle_1.pos.y, particle_1.pos.z) {
-            let (x, y, z) = material.geometry.closest_point(particle_1.pos.x, particle_1.pos.y, particle_1.pos.z);
-            let distance_to = ((x - particle_1.pos.x).powi(2) + (y - particle_1.pos.y).powi(2) + (z - particle_1.pos.z).powi(2)).sqrt();
-            mfp += distance_to;
-        }
 
         for k in 0..(options.weak_collision_order + 1) {
             binary_collision_geometries.push(BinaryCollisionGeometry::new(phis_azimuthal[k], impact_parameters[k], mfp))
