### 2021-03-16 4653945
Cargo.toml update to keep up with semantic versioning.
