### 2021-03-16 5e7852a
Added some tests of the parry3d geometry.
