+++ src/tests.rs
@@ -4,6 +4,193 @@ use super::*;
 use float_cmp::*;
 
 #[test]
+#[cfg(feature = "parry3d")]
+fn test_parry_cuboid() {
+    let mass = 1.;
+    let Z = 1.;
+    let E = 10.*EV;
+    let Ec = 1.*EV;
+    let Es = 5.76*EV;
+    let x = 0.;
+    let y = 0.;
+    let z = 0.;
+    let cosx = 1./(2.0_f64).sqrt();
+    let cosy = 1./(2.0_f64).sqrt();
+    let cosz = 0.;
+    let particle_1 = particle::Particle::new(mass, Z, E, Ec, Es, x, y, z, cosx, cosy, cosz, false, false, 0);
+
+    let material_parameters = material::MaterialParameters{
+        energy_unit: "EV".to_string(),
+        mass_unit: "AMU".to_string(),
+        Eb: vec![0.0, 0.0],
+        Es: vec![2.0, 4.0],
+        Ec: vec![1.0, 1.0],
+        Z: vec![29., 1.],
+        m: vec![63.54, 1.0008],
+        interaction_index: vec![0, 0],
+        surface_binding_model: SurfaceBindingModel::PLANAR{calculation: SurfaceBindingCalculation::TARGET},
+        bulk_binding_model: BulkBindingModel::INDIVIDUAL,
+    };
+
+    let radius = 1000.*ANGSTROM;
+
+    //cuboid centered at (0, 0, 0) with l=w=h=1 Angstrom
+    let geometry_input = parry::ParryTriMeshInput {
+        length_unit: "ANGSTROM".to_string(),
+        densities: vec![0.03, 0.03],
+        electronic_stopping_correction_factor: 1.0,
+        vertices: vec![
+            [-0.5, -0.5, 0.5],
+            [0.5, -0.5, 0.5],
+            [-0.5, 0.5, 0.5],
+            [0.5, 0.5, 0.5],
+            [-0.5, 0.5, -0.5],
+            [0.5, 0.5, -0.5],
+            [-0.5, -0.5, -0.5],
+            [0.5, -0.5, -0.5],
+        ],
+        indices: vec![
+            [0, 1, 3],
+            [0, 3, 2],
+            [2, 3, 5],
+            [2, 5, 4],
+            [4, 5, 7],
+            [4, 7, 6],
+            [6, 7, 1],
+            [6, 1, 0],
+            [1, 7, 5],
+            [1, 5, 3],
+            [6, 0, 2],
+            [6, 2, 4]
+        ]
+    };
+
+    let mut material_cuboid: material::Material<parry::ParryTriMesh> = material::Material::<parry::ParryTriMesh>::new(&material_parameters, &geometry_input);
+    material_cuboid.geometry.energy_barrier_thickness = 10.*ANGSTROM;
+
+    let surface_binding_energy = material_cuboid.actual_surface_binding_energy(&particle_1, particle_1.pos.x, particle_1.pos.y, particle_1.pos.z);
+    assert!(approx_eq!(f64, surface_binding_energy/EV, (2. + 4.)/2., epsilon=1E-24));
+
+    //Test origin is inside
+    assert!(material_cuboid.geometry.inside(0., 0., 0.));
+
+    //Test all outside areas for containment
+    assert!(!material_cuboid.geometry.inside(0., 1., 0.));
+    assert!(!material_cuboid.geometry.inside(0., -1., 0.));
+    assert!(!material_cuboid.geometry.inside(1., 0., 0.));
[…]
