+++ src/main.rs
@@ -60,7 +60,7 @@ pub mod consts;
 pub mod structs;
 pub mod sphere;
 
-#[cfg(feature = "parry")]
+#[cfg(feature = "parry3d")]
 pub mod parry;
 
 pub use crate::enums::*;
@@ -71,7 +71,7 @@ pub use crate::output::{OutputUnits};
 pub use crate::geometry::{Geometry, GeometryElement, Mesh0D, Mesh1D, Mesh2D};
 pub use crate::sphere::{Sphere, SphereInput, InputSphere};
 
-#[cfg(feature = "parry")]
+#[cfg(feature = "parry3d")]
 pub use crate::parry::{ParryBall, ParryBallInput, InputParryBall, ParryTriMesh, ParryTriMeshInput, InputParryTriMesh};
 
 fn physics_loop<T: Geometry + Sync>(particle_input_array: Vec<particle::ParticleInput>, material: material::Material<T>, options: Options, output_units: OutputUnits) {
@@ -164,9 +164,9 @@ fn main() {
             "1D" => GeometryType::MESH1D,
             "2D" => GeometryType::MESH2D,
             "SPHERE" => GeometryType::SPHERE,
-            #[cfg(feature = "parry")]
+            #[cfg(feature = "parry3d")]
             "BALL" => GeometryType::BALL,
-            #[cfg(feature = "parry")]
+            #[cfg(feature = "parry3d")]
             "TRIMESH" => GeometryType::TRIMESH,
             _ => panic!("Unimplemented geometry {}.", args[1].clone())
         }),
@@ -190,12 +190,12 @@ fn main() {
             let (particle_input_array, material, options, output_units) = input::input::<Sphere>(input_file);
             physics_loop::<Sphere>(particle_input_array, material, options, output_units);
         },
-        #[cfg(feature = "parry")]
+        #[cfg(feature = "parry3d")]
         GeometryType::BALL => {
             let (particle_input_array, material, options, output_units) = input::input::<ParryBall>(input_file);
             physics_loop::<ParryBall>(particle_input_array, material, options, output_units);
         }
-        #[cfg(feature = "parry")]
+        #[cfg(feature = "parry3d")]
         GeometryType::TRIMESH => {
             let (particle_input_array, material, options, output_units) = input::input::<ParryTriMesh>(input_file);
             physics_loop::<ParryTriMesh>(particle_input_array, material, options, output_units);
