+++ src/enums.rs
@@ -5,9 +5,9 @@ pub enum MaterialType {
     MESH1D(material::Material<geometry::Mesh1D>),
     MESH2D(material::Material<geometry::Mesh2D>),
     SPHERE(material::Material<sphere::Sphere>),
-    #[cfg(feature = "parry")]
+    #[cfg(feature = "parry3d")]
     BALL(material::Material<parry::ParryBall>),
-    #[cfg(feature = "parry")]
+    #[cfg(feature = "parry3d")]
     TRIMESH(material::Material<parry::ParryTriMesh>)
 }
 
@@ -17,9 +17,9 @@ pub enum GeometryType {
     MESH1D,
     MESH2D,
     SPHERE,
-    #[cfg(feature = "parry")]
+    #[cfg(feature = "parry3d")]
     BALL,
-    #[cfg(feature = "parry")]
+    #[cfg(feature = "parry3d")]
     TRIMESH,
 }
 
