+++ Cargo.toml
@@ -40,5 +40,5 @@ cpr_rootfinder_netlib = ["rcpr", "netlib-src"]
 cpr_rootfinder_intel_mkl = ["rcpr", "intel-mkl-src"]
 distributions = ["ndarray"]
 no_list_output = []
-parry = ["parry3d-f64"]
+parry3d = ["parry3d-f64"]
 accelerated_ions = []
