### 2021-03-19 3db7d02
Cleanup of some redundant traits.
