### 2021-03-19 0aa5c46
Removed nearest_to_* functions, and instead wrapped that functionality into the equivalent functions *.:
