+++ .github/workflows/rustbca_compile_check.yml
@@ -2,9 +2,9 @@ name: rustBCA Compile check
 
 on:
   push:
-    branches: [ master ]
+    branches: [ master, dev ]
   pull_request:
-    branches: [ master ]
+    branches: [ master, dev ]
   schedule:
     - cron: "0 0 * * *"
 
