+++ src/lib.rs
@@ -1,0 +1,184 @@
+#![allow(unused_variables)]
+#![allow(non_snake_case)]
+#![allow(non_camel_case_types)]
+
+#[cfg(feature = "cpr_rootfinder_openblas")]
+extern crate openblas_src;
+#[cfg(feature = "cpr_rootfinder_netlib")]
+extern crate netlib_src;
+#[cfg(feature = "cpr_rootfinder_intel_mkl")]
+extern crate intel_mkl_src;
+
+use std::{fmt};
+use std::mem::discriminant;
+
+//Error handling crate
+use anyhow::Result;
+use anyhow::*;
+
+//Serializing/Deserializing crate
+use serde::*;
+
+//Array input via hdf5
+#[cfg(feature = "hdf5_input")]
+use hdf5::*;
+
+//I/O
+use std::fs::OpenOptions;
+use std::io::prelude::*;
+use std::io::BufWriter;
+
+//itertools
+use itertools::izip;
+
+//Math
+use std::f64::consts::FRAC_2_SQRT_PI;
+use std::f64::consts::PI;
+use std::f64::consts::SQRT_2;
+
+use pyo3::prelude::*;
+use pyo3::wrap_pyfunction;
+
+//Load internal modules
+pub mod material;
+pub mod particle;
+pub mod tests;
+pub mod interactions;
+pub mod bca;
+pub mod geometry;
+pub mod input;
+pub mod output;
+pub mod enums;
+pub mod consts;
+pub mod structs;
+pub mod sphere;
+
+#[cfg(feature = "parry3d")]
+pub mod parry;
+
+pub use crate::enums::*;
+pub use crate::consts::*;
+pub use crate::structs::*;
+pub use crate::input::{Input2D, Input1D, Input0D, Options, InputFile, GeometryInput};
+pub use crate::output::{OutputUnits};
+pub use crate::geometry::{Geometry, GeometryElement, Mesh0D, Mesh1D, Mesh2D};
+pub use crate::sphere::{Sphere, SphereInput, InputSphere};
+
+#[cfg(feature = "parry3d")]
+pub use crate::parry::{ParryBall, ParryBallInput, InputParryBall, ParryTriMesh, ParryTriMeshInput, InputParryTriMesh};
+
+#[pymodule]
+pub fn pybca(py: Python, m: &PyModule) -> PyResult<()> {
+    m.add_function(wrap_pyfunction!(simple_bca_py, m)?)?;
+    Ok(())
+}
+
+#[repr(C)]
+pub struct OutputBCA {
+    len: usize,
+    pub particles: *mut [f64; 9],
+}
+
+#[no_mangle]
+pub extern "C" fn simple_bca_c(x: f64, y: f64, z: f64, ux: f64, uy: f64, uz: f64, E1: f64, Z1: f64, m1: f64, Ec1: f64, Es1: f64, Z2: f64, m2: f64, Ec2: f64, Es2: f64, n2: f64, Eb2: f64) -> OutputBCA {
+    let mut output = simple_bca(x, y, z, ux, uy, uz, E1, Z1, m1, Ec1, Es1, Z2, m2, Ec2, Es2, n2, Eb2);
+    let len = output.len();
+    let particles = output.as_mut_ptr();
+
+    std::mem::forget(output);
+    OutputBCA {
+        len,
+        particles
+    }
+}
+
+#[pyfunction]
[…]
