### 2021-03-19 b0a899f
Added unneeded but necessary-to-compile options.
