### 2021-03-20 f2b7dc7
Added ability to run lists of energy-angle coords for simple_bca.
