### 2021-03-23 8511c65
First pass of library error messages.
