### 2021-03-26 2811710
Added context to toml unpack that hints to the user they may have forgotten the geometry command line argument (0D, 1D, 2D, SPHERE, BALL, TRIMESH)
