+++ src/input.rs
@@ -28,7 +28,7 @@ impl GeometryInput for Input2D {
 impl InputFile for Input2D {
 
     fn new(string: &str) -> Input2D {
-        toml::from_str(string).unwrap()
+        toml::from_str(string).context("Could not parse TOML file. Be sure you are using the correct input file mode (e.g., ./RustBCA SPHERE sphere.toml or RustBCA.exe 0D mesh_0d.toml).").unwrap()
     }
 
     fn get_options(&self) -> &Options{
@@ -60,7 +60,7 @@ impl GeometryInput for Input0D {
 impl InputFile for Input0D {
 
     fn new(string: &str) -> Input0D {
-        toml::from_str(string).expect("Could not parse TOML file.")
+        toml::from_str(string).context("Could not parse TOML file. Be sure you are using the correct input file mode (e.g., ./RustBCA SPHERE sphere.toml or RustBCA.exe 0D mesh_0d.toml).").unwrap()
     }
 
     fn get_options(&self) -> &Options{
@@ -92,7 +92,7 @@ impl GeometryInput for Input1D {
 impl InputFile for Input1D {
 
     fn new(string: &str) -> Input1D {
-        toml::from_str(string).expect("Could not parse TOML file.")
+        toml::from_str(string).context("Could not parse TOML file. Be sure you are using the correct input file mode (e.g., ./RustBCA SPHERE sphere.toml or RustBCA.exe 0D mesh_0d.toml).").unwrap()
     }
 
     fn get_options(&self) -> &Options{
