+++ src/sphere.rs
@@ -11,7 +11,7 @@ pub struct InputSphere {
 impl InputFile for InputSphere {
 
     fn new(string: &str) -> InputSphere {
-        toml::from_str(string).expect("Could not parse TOML file.")
+        toml::from_str(string).context("Could not parse TOML file. Be sure you are using the correct input file mode (e.g., ./RustBCA SPHERE sphere.toml or RustBCA.exe 0D mesh_0d.toml).").unwrap()
     }
 
     fn get_options(&self) -> &Options{
