### 2021-03-26 5d81514
Progress bar .finish() call was not in the right place.
