+++ src/lib.rs
@@ -28,6 +28,9 @@ use std::fs::OpenOptions;
 use std::io::prelude::*;
 use std::io::BufWriter;
 
+//standard slice
+use std::slice;
+
 //itertools
 use itertools::izip;
 
@@ -76,13 +79,192 @@ pub fn pybca(py: Python, m: &PyModule) -> PyResult<()> {
 
 #[repr(C)]
 pub struct OutputBCA {
-    len: usize,
+    pub len: usize,
     pub particles: *mut [f64; 9],
+}
+
+#[derive(Debug)]
+#[repr(C)]
+pub struct InputSimpleBCA {
+    pub len: usize,
+    /// x, y, z, vx, vy, vz
+    pub velocities: *mut [f64; 6],
+    pub Z1: f64,
+    pub m1: f64,
+    pub Ec1: f64,
+    pub Es1: f64,
+    pub Z2: f64,
+    pub m2: f64,
+    pub n2: f64,
+    pub Ec2: f64,
+    pub Es2: f64,
+    pub Eb2: f64,
+}
+
+#[no_mangle]
+pub extern "C" fn simple_bca_list_c(input: InputSimpleBCA) -> OutputBCA {
+
+    let x = -2.*(input.n2*10E30).powf(-1./3.);
+    let y = 0.0;
+    let z = 0.0;
+
+    let mut total_output = vec![];
+
+    #[cfg(feature = "distributions")]
+    let options = Options {
+        name: "test".to_string(),
+        track_trajectories: false,
+        track_recoils: true,
+        track_recoil_trajectories: false,
+        write_buffer_size: 8000,
+        weak_collision_order: 3,
+        suppress_deep_recoils: false,
+        high_energy_free_flight_paths: false,
+        electronic_stopping_mode: ElectronicStoppingMode::INTERPOLATED,
+        mean_free_path_model: MeanFreePathModel::LIQUID,
+        interaction_potential: vec![vec![InteractionPotential::KR_C]],
+        scattering_integral: vec![vec![ScatteringIntegral::MENDENHALL_WELLER]],
+        num_threads: 1,
+        num_chunks: 1,
+        use_hdf5: false,
+        root_finder: vec![vec![Rootfinder::DEFAULTNEWTON]],
+        track_displacements: false,
+        track_energy_losses: false,
+        energy_min: 0.0,
+        energy_max: 10.0,
+        energy_num: 11,
+        angle_min: 0.0,
+        angle_max: 90.0,
+        angle_num: 11,
+        x_min: 0.0,
+        y_min: -10.0,
+        z_min: -10.0,
+        x_max: 10.0,
+        y_max: 10.0,
+        z_max: 10.0,
+        x_num: 11,
+        y_num: 11,
+        z_num: 11,
+    };
+
+    #[cfg(not(feature = "distributions"))]
+    let options = Options {
+        name: "test".to_string(),
+        track_trajectories: false,
+        track_recoils: true,
+        track_recoil_trajectories: false,
+        write_buffer_size: 8000,
+        weak_collision_order: 3,
+        suppress_deep_recoils: false,
+        high_energy_free_flight_paths: false,
+        electronic_stopping_mode: ElectronicStoppingMode::INTERPOLATED,
+        mean_free_path_model: MeanFreePathModel::LIQUID,
+        interaction_potential: vec![vec![InteractionPotential::KR_C]],
+        scattering_integral: vec![vec![ScatteringIntegral::MENDENHALL_WELLER]],
+        num_threads: 1,
+        num_chunks: 1,
+        use_hdf5: false,
+        root_finder: vec![vec![Rootfinder::DEFAULTNEWTON]],
+        track_displacements: false,
+        track_energy_losses: false,
+    };
+
[…]
