### 2021-03-29 347bc24
Updates to C bindings.
