### 2021-03-30 7634de3
Fix to filter out stopped, non-incident ions.
