### 2021-05-21 #3101 @arfon
:wave: @HaoZeke @joglekara @amritagos – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission we'd be asking you to review here is [_RustBCA: A High-Performance, Binary-Collision-Approximation Code for Ion-Material Interactions_](https://github.com/lcpp-org/RustBCA).
