### 2021-06-11 322aad7
Added package.py for spack compatibility.
