### 2021-06-18 lcpp-org/RustBCA#134 @joglekara
[JOSS Review] Python requirements file
re: https://github.com/openjournals/joss-reviews/issues/3298

It will be useful to add a requirements file for the Pythonic post-processing performed here. e.g. `shapely`, `toml`
