### 2021-07-02 lcpp-org/RustBCA#136 @joglekara
[JOSS Review] - File name of examples does not match those in manuscript
**Description**
File name of examples does not match those in manuscript

**To Reproduce**
cargo run --release examples/layered_target.toml
cargo run --release examples/layered_target_1D.toml

**Expected behavior**
A clear and concise description of what you expected to happen.
Run the example

**Error messages, output files, or figures**
Please include error messages, output files or figures that demonstrate the error if applicable.
`thread 'main' panicked at 'Input errror: could not open input file examples/layered_target.toml.: Os { code: 2, kind: NotFound, message: "No such file or directory" }', src/input.rs:184:10`

**System (please complete the following information):**
OS:

**Additional context**
I am not clear on which filename would be preferable so leaving it up to @drobnyjt to determine if the change needs to be in the manuscript or in the example file name
