### 2021-07-22 bb36666
Minor updates to scripts.
