+++ scripts/rustbca.py
@@ -10,8 +10,8 @@ from matplotlib import rcParams, cm
 import matplotlib as mpl
 import matplotlib.colors as colors
 
-from .materials import *
-from .formulas import *
+from materials import *
+from formulas import *
 
 #from generate_ftridyn_input import *
 
@@ -170,7 +170,7 @@ def do_trajectory_plot_3d(name, thickness=None, depth=None, boundary=None, plot_
     index = 0
     x_max = 0
     min_length = 1
-    scale_factor = 200.0
+    scale_factor = 1.0
     if np.size(trajectories) > 0:
         for trajectory_length in trajectory_data:
 
@@ -276,9 +276,9 @@ def generate_rustbca_input(Zb, Mb, n, Eca, Ecb, Esa, Esb, Eb, Ma, Za, E0, N, N_,
     track_displacements=True, track_energy_losses=True,
     electronic_stopping_mode=LOW_ENERGY_NONLOCAL,
     weak_collision_order=3, ck=1., mean_free_path_model=LIQUID,
-    interaction_potential='"KR_C"', high_energy=False, energy_barrier_thickness=(6.306E10)**(-1/3.),
-    initial_particle_position = -1*ANGSTROM/MICRON, integral='"MENDENHALL_WELLER"',
-    root_finder = '{"NEWTON"={max_iterations=100, tolerance=1e-3}}',
+    interaction_potential="KR_C", high_energy=False, energy_barrier_thickness=(6.306E10)**(-1/3.),
+    initial_particle_position = -1*ANGSTROM/MICRON, integral="MENDENHALL_WELLER",
+    root_finder = "DEFAULTNEWTON",
     delta_x_angstrom=5., uniformly_distributed_ions=False):
 
     '''
@@ -295,7 +295,7 @@ def generate_rustbca_input(Zb, Mb, n, Eca, Ecb, Esa, Esb, Eb, Ma, Za, E0, N, N_,
         'suppress_deep_recoils': False,
         'high_energy_free_flight_paths': high_energy,
         'num_threads': 4,
-        'num_chunks': 100,
+        'num_chunks': 10,
         'use_hdf5': False,
         'electronic_stopping_mode': electronic_stopping_mode,
         'mean_free_path_model': mean_free_path_model,
@@ -374,7 +374,7 @@ def plot_distributions_rustbca(name, beam, target,
     incident_energy=1, incident_angle=0,
     max_collision_contours=4, plot_2d_reflected_contours=False,
     collision_contour_significance_threshold=0.1, plot_garrison_contours=False,
-    plot_reflected_energies_by_number_collisions=False,
+    plot_reflected_energies_by_number_collisions=True,
     plot_scattering_energy_curve=False):
 
     '''
@@ -389,7 +389,7 @@ def plot_distributions_rustbca(name, beam, target,
         max_collision_contours (int): number of collision event contours to plot on reflected EAD; default 4
         plot_2d_reflected_contours (bool): whether to plot collision event contours on reflected EAD; default False
         collision_contour_significance_threshold (int): threshold of significance for contour plotting; default 0.1
-        plot_reflected_energies_by_number_collisions (bool): separte reflected energy spectrum into collision number distributions; default False
+        RustBCA.exe (bool): separte reflected energy spectrum into collision number distributions; default False
         plot_scattering_energy_curve (bool): plot bold, translucent white curve showing theoretical single-collision reflected energies; default False
 
     '''
@@ -866,7 +866,7 @@ def run_iead(ions, target, energies, angles, iead, name="default_", N=1):
[…]
