### 2021-07-22 e2402fe
Update package.py
Minor update to improve Spack compatibility.
