+++ src/lib.rs
@@ -101,6 +101,25 @@ pub struct InputSimpleBCA {
     pub Eb2: f64,
 }
 
+#[derive(Debug)]
+#[repr(C)]
+pub struct InputCompoundBCA {
+    pub len: usize,
+    /// vx, vy, vz
+    pub velocities: *mut [f64; 3],
+    pub Z1: f64,
+    pub m1: f64,
+    pub Ec1: f64,
+    pub Es1: f64,
+    pub num_species_target: usize,
+    pub Z2: *mut f64,
+    pub m2: *mut f64,
+    pub n2: *mut f64,
+    pub Ec2: *mut f64,
+    pub Es2: *mut f64,
+    pub Eb2: *mut f64,
+}
+
 #[no_mangle]
 pub extern "C" fn simple_bca_list_c(input: InputSimpleBCA) -> OutputBCA {
 
@@ -185,6 +204,176 @@ pub extern "C" fn simple_bca_list_c(input: InputSimpleBCA) -> OutputBCA {
     let geometry_input = geometry::Mesh0DInput {
         length_unit: "ANGSTROM".to_string(),
         densities: vec![input.n2],
+        electronic_stopping_correction_factor: 1.0
+    };
+
+    let m = material::Material::<Mesh0D>::new(&material_parameters, &geometry_input);
+
+    let velocities = unsafe { slice::from_raw_parts(input.velocities, input.len) };
+
+    for velocity in velocities {
+
+        let vx = velocity[0];
+        let vy = velocity[1];
+        let vz = velocity[2];
+
+        let v = (vx*vx + vy*vy + vz*vz).sqrt();
+
+        let E1 = 0.5*input.m1*AMU*v*v;
+
+        let ux = vx/v;
+        let uy = vy/v;
+        let uz = vz/v;
+
+        let p = particle::Particle {
+            m: input.m1*AMU,
+            Z: input.Z1,
+            E: E1,
+            Ec: input.Ec1*EV,
+            Es: input.Es1*EV,
+            pos: Vector::new(x, y, z),
+            dir: Vector::new(ux, uy, uz),
+            pos_origin: Vector::new(x, y, z),
+            pos_old: Vector::new(x, y, z),
+            dir_old: Vector::new(ux, uy, uz),
+            energy_origin: E1,
+            asymptotic_deflection: 0.0,
+            stopped: false,
+            left: false,
+            incident: true,
+            first_step: true,
+            trajectory: vec![],
+            energies: vec![],
+            track_trajectories: false,
+            number_collision_events: 0,
+            backreflected: false,
+            interaction_index : 0
+        };
+
+
+        let output = bca::single_ion_bca(p, &m, &options);
+
+        for particle in output {
+
+            if (particle.left) | (particle.incident) {
+                total_output.push(
+                    [
+                        particle.Z,
+                        particle.m/AMU,
+                        particle.E/EV,
+                        particle.pos.x/ANGSTROM,
+                        particle.pos.y/ANGSTROM,
+                        particle.pos.z/ANGSTROM,
+                        particle.dir.x,
+                        particle.dir.y,
+                        particle.dir.z
+                    ]
+                );
+            }
+        }
+    }
+
+    let len = total_output.len();
+    let particles = total_output.as_mut_ptr();
+
+    std::mem::forget(total_output);
+    OutputBCA {
+        len,
+        particles
+    }
+}
+
+#[no_mangle]
+pub extern "C" fn compound_bca_list_c(input: InputCompoundBCA) -> OutputBCA {
+
+    let mut total_output = vec![];
+
+    #[cfg(feature = "distributions")]
+    let options = Options {
[…]
