### 2021-08-03 lcpp-org/RustBCA#143 @drobnyjt
I don't have easy access to a Mac, but will start looking into this. It appears to be a problem with PyO3, the crate used to generate the Python module from lib.rs. Since the Python bindings should not be necessary to simply run the examples, I'll see what's necessary to get cargo run to build the binary only.
