+++ Cargo.toml
@@ -32,6 +32,7 @@ parry3d-f64 = {version = "0.2.0", optional = true}
 [dependencies.pyo3]
 version = "0.13.2"
 features = ["extension-module"]
+optional = true
 
 [dev-dependencies]
 float-cmp = "0.8.0"
@@ -51,3 +52,4 @@ distributions = ["ndarray"]
 no_list_output = []
 parry3d = ["parry3d-f64"]
 accelerated_ions = []
+python = ["pyo3"]
