### 2021-08-03 8a4d96e
Updated setup.py to ensure pyo3 is built.
