### 2021-08-03 64bcac7
More fixes to setup.py - I was using the setuptools Rust API incorrectly, and it was silently failing.
