+++ setup.py
@@ -3,8 +3,7 @@ from setuptools_rust import Binding, RustExtension
 
 setup(
     name="RustBCA",
-    rust_extensions=[RustExtension("libRustBCA.pybca", binding=Binding.PyO3)],
+    rust_extensions=[RustExtension("libRustBCA.pybca", binding=Binding.PyO3, features=["python"])],
     # rust extensions are not zip safe, just like C-extensions.
-    features=["python"],
     zip_safe=False,
 )
