### 2021-08-03 35e3ea1
Add python binding installation to workflow
I've added a simple check to make sure the python bindings install correctly from scratch.
