+++ .github/workflows/rustbca_compile_check.yml
@@ -37,7 +37,7 @@ jobs:
         sudo apt install libhdf5-dev
     - name: test Python Bindings
       run: |
-        suto python3 -m pip install setuptools_rust testresources
+        sudo python3 -m pip install setuptools_rust testresources
         sudo python3 setup.py install
         python3 -c "from libRustBCA.pybca import *; print(simple_bca_py)"
     - name: Test RustBCA
