### 2021-08-03 0943871
Ensure rust compiler is installed
Python was having trouble finding the rust compiler, so I've added that to the workflow.
