+++ .github/workflows/rustbca_compile_check.yml
@@ -21,6 +21,7 @@ jobs:
     - name: Install rust
       run: |
         curl --proto '=https' --tlsv1.2 -sSf -y https://sh.rustup.rs | sh
+        sudo apt-get install rustc
     - name: Install pip for Python-3
       run: |
         sudo apt-get install python3-pip
