### 2021-08-03 8dbc9c5
Add sudo
For some reason, cargo test is now throwing a permission denied error in the workflow - sudo added to see if that fixes it.
