+++ .github/workflows/rustbca_compile_check.yml
@@ -47,12 +47,12 @@ jobs:
     - name: Run Examples
       run: |
         sudo cargo run --release 0D examples/boron_nitride_0D.toml
-        ./target/release/RustBCA 0D examples/titanium_dioxide_0D.toml
-        ./target/release/RustBCA 1D examples/layered_geometry_1D.toml
+        sudo ./target/release/RustBCA 0D examples/titanium_dioxide_0D.toml
+        sudo ./target/release/RustBCA 1D examples/layered_geometry_1D.toml
         cat 2000.0eV_0.0001deg_He_TiO2_Al_Sisummary.output
-        ./target/release/RustBCA examples/boron_nitride.toml
-        ./target/release/RustBCA examples/layered_geometry.toml
+        sudo ./target/release/RustBCA examples/boron_nitride.toml
+        sudo ./target/release/RustBCA examples/layered_geometry.toml
         cat 2000.0eV_0.0001deg_He_TiO2_Al_Sisummary.output
-        ./target/release/RustBCA SPHERE examples/boron_nitride_sphere.toml
+        sudo ./target/release/RustBCA SPHERE examples/boron_nitride_sphere.toml
         sudo cargo run --release --features parry3d TRIMESH examples/tungsten_twist_trimesh.toml
         
