### 2021-08-21 #3298 @HaoZeke
The authors have made a sincere effort towards fixing the MacOS installation; and I do not believe it is a blocker. The rest of the work stands nicely as is. LGTM.
