+++ src/input.rs
@@ -395,7 +395,9 @@ where <T as Geometry>::InputFileFormat: Deserialize<'static> + 'static {
                             ux: cosx,
                             uy: cosy,
                             uz: cosz,
-                            interaction_index
+                            interaction_index,
+                            weight: 1.0,
+                            tag: 0,
                         }
                     );
                 }
