+++ src/lib.rs
@@ -123,6 +123,226 @@ pub struct InputCompoundBCA {
     pub Eb2: *mut f64,
 }
 
+#[derive(Debug)]
+#[repr(C)]
+pub struct InputTaggedBCA {
+    pub len: usize,
+    /// x y z
+    pub positions: *mut [f64; 3],
+    /// vx, vy, vz
+    pub velocities: *mut [f64; 3],
+    pub Z1: f64,
+    pub m1: f64,
+    pub Ec1: f64,
+    pub Es1: f64,
+    pub num_species_target: usize,
+    pub Z2: *mut f64,
+    pub m2: *mut f64,
+    pub n2: *mut f64,
+    pub Ec2: *mut f64,
+    pub Es2: *mut f64,
+    pub Eb2: *mut f64,
+    pub tags: *mut i32,
+    pub weights: *mut f64,
+}
+
+#[derive(Debug)]
+#[repr(C)]
+pub struct OutputTaggedBCA {
+    pub len: usize,
+    pub particles: *mut [f64; 9],
+    pub weights: *mut f64,
+    pub tags: *mut i32
+}
+
+#[no_mangle]
+pub extern "C" fn compound_tagged_bca_list_c(input: InputTaggedBCA) -> OutputTaggedBCA {
+
+    let mut total_output = vec![];
+    let mut output_tags = vec![];
+    let mut output_weights = vec![];
+
+    #[cfg(feature = "distributions")]
+    let options = Options {
+        name: "test".to_string(),
+        track_trajectories: false,
+        track_recoils: true,
+        track_recoil_trajectories: false,
+        write_buffer_size: 8000,
+        weak_collision_order: 3,
+        suppress_deep_recoils: false,
+        high_energy_free_flight_paths: false,
+        electronic_stopping_mode: ElectronicStoppingMode::INTERPOLATED,
+        mean_free_path_model: MeanFreePathModel::LIQUID,
+        interaction_potential: vec![vec![InteractionPotential::KR_C]],
+        scattering_integral: vec![vec![ScatteringIntegral::MENDENHALL_WELLER]],
+        num_threads: 1,
+        num_chunks: 1,
+        use_hdf5: false,
+        root_finder: vec![vec![Rootfinder::DEFAULTNEWTON]],
+        track_displacements: false,
+        track_energy_losses: false,
+        energy_min: 0.0,
+        energy_max: 10.0,
+        energy_num: 11,
+        angle_min: 0.0,
+        angle_max: 90.0,
+        angle_num: 11,
+        x_min: 0.0,
+        y_min: -10.0,
+        z_min: -10.0,
+        x_max: 10.0,
+        y_max: 10.0,
+        z_max: 10.0,
+        x_num: 11,
+        y_num: 11,
+        z_num: 11,
+    };
+
+    #[cfg(not(feature = "distributions"))]
+    let options = Options {
+        name: "test".to_string(),
+        track_trajectories: false,
+        track_recoils: true,
+        track_recoil_trajectories: false,
+        write_buffer_size: 8000,
+        weak_collision_order: 3,
+        suppress_deep_recoils: false,
+        high_energy_free_flight_paths: false,
+        electronic_stopping_mode: ElectronicStoppingMode::INTERPOLATED,
+        mean_free_path_model: MeanFreePathModel::LIQUID,
+        interaction_potential: vec![vec![InteractionPotential::KR_C]],
+        scattering_integral: vec![vec![ScatteringIntegral::MENDENHALL_WELLER]],
+        num_threads: 1,
+        num_chunks: 1,
+        use_hdf5: false,
+        root_finder: vec![vec![Rootfinder::DEFAULTNEWTON]],
+        track_displacements: false,
[…]
