+++ src/structs.rs
@@ -1,5 +1,5 @@
 /// 3D vector.
-#[derive(Clone)]
+#[derive(Clone, Copy)]
 pub struct Vector {
     pub x: f64,
     pub y: f64,
