+++ src/input.rs
@@ -351,7 +351,10 @@ where <T as Geometry>::InputFileFormat: Deserialize<'static> + 'static {
                             ux: cosx,
                             uy: cosy,
                             uz: cosz,
-                            interaction_index: interaction_index
+                            interaction_index: interaction_index,
+                            tag: 0,
+                            weight: 1.0,
+                            tracked_vector: Vector::new(0.0, 0.0, 0.0),
                         }
                     );
                 }
