### 2021-08-22 bb5c48a
Fixes to input deck - no need to have weight/tag in toml right now.
