### 2021-08-30 #3298 @drobnyjt
@arfon Sure! I have published a new [release ](https://github.com/lcpp-org/RustBCA/releases/tag/v1.1.1-joss)and created an archive on Zenodo (https://doi.org/10.5281/zenodo.5337083).
